+++ CorePayments/src/main/java/com/paypal/android/corepayments/CoreConfig.kt
@@ -3,6 +3,6 @@ package com.paypal.android.corepayments
 data class CoreConfig @JvmOverloads constructor(
     val clientId: String,
     val merchantId: String,
-    val environment: Environment = Environment.SANDBOX,
+    val coreEnvironment: CoreEnvironment = CoreEnvironment.SANDBOX,
     val bnCode: String? = null,
 )