+++ PaymentButtons/src/main/java/com/paypal/android/paymentbuttons/PaymentButton.kt
@@ -18,7 +18,7 @@ import com.google.android.material.shape.MaterialShapeDrawable
 import com.google.android.material.shape.RoundedCornerTreatment
 import com.google.android.material.shape.ShapeAppearanceModel
 import com.paypal.android.corepayments.CoreConfig
-import com.paypal.android.corepayments.Environment
+import com.paypal.android.corepayments.CoreEnvironment
 import com.paypal.android.corepayments.analytics.AnalyticsEventData
 import com.paypal.android.corepayments.analytics.AnalyticsService
 import com.paypal.android.ui.R
@@ -36,7 +36,7 @@ abstract class PaymentButton<C : PaymentButtonColor> @JvmOverloads constructor(
     private var shapeHasChanged = false
 
     internal val analyticsService: AnalyticsService =
-        AnalyticsService(context, CoreConfig(clientId = "N/A", merchantId = "N/A", environment = Environment.LIVE))
+        AnalyticsService(context, CoreConfig(clientId = "N/A", merchantId = "N/A", coreEnvironment = CoreEnvironment.LIVE))
 
     private var shapeAppearanceModel: ShapeAppearanceModel = ShapeAppearanceModel()
         set(value) {